refact: Enrollment 클래스를 만들면서 사용하지 않는 클래스 및 테스트 삭제
