+++ src/main/java/nextstep/courses/domain/session/Session.java
@@ -2,33 +2,56 @@ package nextstep.courses.domain.session;
 
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUser;
+import org.springframework.util.Assert;
 
 import java.time.LocalDateTime;
 
 public class Session {
-    private SessionMakingData sessionMakingData;
+    private String title;
+    private SessionDate sessionDate;
+    private CoverImage coverImage;
     private Enrollment enrollment;
 
     public Session(final String title, final long price, final LocalDateTime startDate, final LocalDateTime endDate) {
-        this(new SessionInfo(title, price), new SessionDate(startDate, endDate), null);
+        this(title, price, new SessionDate(startDate, endDate), null);
     }
 
     public Session(final String title, final long price, final LocalDateTime startDate, final LocalDateTime endDate, final CoverImage coverImage) {
-        this(new SessionInfo(title, price), new SessionDate(startDate, endDate), coverImage);
+        this(title, price, new SessionDate(startDate, endDate), coverImage);
     }
 
     public Session(final String title, final LocalDateTime startDate, final LocalDateTime endDate) {
-        this(new SessionInfo(title, 0L), new SessionDate(startDate, endDate), null);
+        this(title, 0L, new SessionDate(startDate, endDate), null);
     }
 
-    public Session(final SessionInfo sessionInfo, final SessionDate sessionDate, CoverImage coverImage) {
-        this.sessionMakingData = new SessionMakingData(sessionInfo, sessionDate, coverImage);
+    public Session(final String title, final long price, final SessionDate sessionDate, CoverImage coverImage) {
+        validateSession(title, sessionDate);
 
-        this.enrollment = new Enrollment(sessionInfo.getPrice());
+        this.title = title;
+        this.enrollment = new Enrollment(price);
+        this.sessionDate = sessionDate;
+        this.coverImage = validateCoverImage(coverImage);
     }
 
-    public long getPrice() {
-        return this.sessionMakingData.getPrice();
+    private void validateSession(final String title, final SessionDate sessionDate) {
+        Assert.hasText(title, "title cannot be blank");
+        Assert.notNull(sessionDate, "session date cannot be null");
+    }
+
+    private CoverImage validateCoverImage(final CoverImage coverImage) {
+        if (coverImage == null) {
+            return CoverImage.defaultCoverImage();
+        }
+
+        return coverImage;
+    }
+
+    public CoverImage getCoverImage() {
+        return this.coverImage;
+    }
+
+    public long getPrice2() {
+        return this.enrollment.getPrice();
     }
 
     public int getCurrentStudentCount2() {
@@ -41,10 +64,6 @@ public class Session {
 
     public boolean isNotRecruiting2() {
         return !enrollment.isRecruiting();
-    }
-
-    public CoverImage getCoverImage() {
-        return this.sessionMakingData.getCoverImage();
     }
 
     public void ready() {
